[#62] :recycle: refactor(DB): book, user_book 모델 파일 분리
