+++ server/src/db/models/book.ts
@@ -64,47 +64,11 @@ class BookModel {
     }
   }
 
-  async getByUserId(userIdDTO: UserIdDTO, columnArr: string[] = ["*"]) {
-    const joinQuery = `join book on book.bookId = user_book.bookId`;
-    const countQuery = `(select count(*) from user_book as ub where ub.bookId = user_book.bookId) as numMembers`;
-    columnArr = ["book.bookId", "bookName", "color", countQuery];
-    const { query, values } = userBookBuildQuery.makeSelectQuery(
-      { ...userIdDTO },
-      columnArr,
-      joinQuery
-    );
-    logger.info(query);
-    logger.debug(values);
-    const [result] = await pool.query<RowDataPacket[]>(query, values);
-    logger.debug(result);
-    return result;
-  }
-
-  async checkUserBook(userBookDTO: UserBookDTO) {
-    const { query, values } = userBookBuildQuery.makeSelectQuery({ ...userBookDTO });
-    const existsQuery = `SELECT EXISTS (${query})`;
-    logger.info(query);
-    logger.debug(values);
-    const [result] = await pool.query<RowDataPacket[]>(existsQuery, values);
-    logger.debug(result);
-    const [isMember] = Object.values(result[0]);
-    return isMember;
-  }
-
   async pacth(whereDTO: GetBookDTO, bookDTO: UpdateBookDTO) {
     const { query, values } = bookBuildQuery.makeUpdateQuery({ ...whereDTO }, { ...bookDTO });
     logger.info(query);
     logger.debug(values);
     const [result] = await pool.query<ResultSetHeader>(query, values);
-    logger.debug(result);
-    return result;
-  }
-
-  async outBookById(userBookDTO: UserBookDTO) {
-    const { query, values } = userBookBuildQuery.makeDeleteQuery({ ...userBookDTO });
-    logger.info(query);
-    logger.debug(values);
-    const [result] = await pool.query<OkPacket>(query, values);
     logger.debug(result);
     return result;
   }
