+++ server/src/db/models/index.ts
@@ -1,5 +1,6 @@
 export * from "./user";
 export * from "./book";
+export * from "./user-book";
 export * from "./diary";
 export * from "./comment";
 export * from "./invtt-code";
